requirements.txt
numpy>=1.26
opencv-python>=4.9
PyYAML>=6.0
ultralytics>=8.2
lap>=0.5.12
matplotlib>=3.8
